Add a simple script to dump word frequencies to CSV
I wrote this to allow us to do some comparison of the terms in each of
our indexes.  It's also useful to see what terms we're actually
generating for building the spelling correction on top of.

This script requires a python environment with the "requests" module
loaded.

It also requires elasticsearch to be running with plenty of heap memory
allocated to it: I've run it successfully with 2Gb.

It also requires the lowercase_word_split_text field to exist, which
we're adding as part of the spelling correction work.