Remove restriction on indexing pages from smart-answers
We originally thought that smart-answers only published the main
sections of a smart-answer (the flows) which we don't want to index in
rummager, but it also publishes the start pages as well, which we _do_
want to index.

Start pages have a document type of `transaction` and flows have a
document type of `smart_answer`, so the existing list of indexable types
will ensure that only the smart answers are indexed. We don't need to
check the publishing app as well.