Delete unnecessary test logic for smart answers
Delete code which prevented the random schema generator from generating
smart answers in tests. This limitation was originally added because
rummager did not index pages published by smart-answers, so it was
incorrect for tests to use smart-answers as a schema example.

But now we are indexing some documents from smart-answers (the start
pages), so this limitation can be removed from the tests.